factor out gpu flow problem
